Fix Rating

+++ appmovilandroid/app/src/main/java/mx/itesm/segi/perfectproject/OtherProfileFrag.java
@@ -3,15 +3,11 @@ package mx.itesm.segi.perfectproject;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageButton;
 import android.widget.ImageView;
-import android.widget.LinearLayout;
 import android.widget.RatingBar;
 
 import java.util.ArrayList;
@@ -69,7 +65,7 @@ public class OtherProfileFrag extends Fragment {
         tvName.setText(user.getName());
         tvCompany.setText(user.getCompany());
         rateNum.setText(user.getRating() + "");
-        rbRating.setRating(user.getRating());
+        rbRating.setRating((float) user.getRating());
 
         if(getArguments().getBoolean(ARG_JOINED)) tvEmail.setText(user.getEmail());
         else tvEmail.setText(R.string.email_not_available);
